add useTheme hook
